zero warnings build + found a real dangling pointer bug doing it
cleaned up the strncpy truncation warnings (KDB_STRLCPY now uses snprintf,
same deal but gcc stops crying about it), bumped a couple path buffers so
the .tmp/.lock suffix cant theoretically overflow, fixed a bad %lld format
in bench.c, and wired the orphaned storage test into the suite.

while chasing these down -Wdangling-pointer caught something actually
bad in release builds: kdb_list_tables was returning pointers into a
stack-local array that died the moment the function returned. worked by
accident because nothing clobbered the stack before you read it, but it
was UB waiting to bite someone. switched it to a thread-local buffer,
same pattern kdb_last_error already uses, and documented the lifetime
on the declaration

+++ tools/bench.c
@@ -170,7 +170,7 @@ int main(int argc, char **argv) {
     bench_delete     (db, n);
     bench_compact    (db);
 
-    printf("\nFinal row count: %lld\n", kdb_count(db, TABLE_NAME, NULL));
+    printf("\nFinal row count: %lld\n", (long long)kdb_count(db, TABLE_NAME, NULL));
 
     kdb_close(db);
     return 0;
